requirements.txt
streamlit
openai
pycountry
load_dotenv()